backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
sqlalchemy==2.0.35
pypdf==6.1.0
python-docx==1.1.2
openai>=1.50.0
httpx==0.27.2
python-dotenv==1.0.1
aiofiles==23.2.1
